refactor(server-web): promote Sources/Discovery/Mapping/Resolved out of Settings
The previous shape parked Sources / Discovery / Mapping / Resolved
inside a "Settings" page as tabs. That was honest for General + danger
zone but Discovery and Mapping are full-screen workspaces, not
"settings" — and the 2779-line monolith proved the category was a
wastebasket.

Restructured under `topologies/[id]/`:

  /topologies/[id]                    ← Diagram (default workspace)
  /topologies/[id]/sources            ← what's attached, merge config
  /topologies/[id]/discovery          ← identity binding + policy + sync
  /topologies/[id]/mapping            ← node/link → host/interface
  /topologies/[id]/resolved           ← debug JSON viewer
  /topologies/[id]/settings           ← display + danger zone only

Each is a real subroute (Ctrl-click + back/forward + deep links all
work). Tab bar lives in `topologies/[id]/+layout.svelte` and is hidden
on the diagram route so the canvas keeps full bleed.

Shared state (topology, sources, data-source catalogs) loads once in
the layout and lives in `_context.svelte.ts`, consumed by children
via `useTopologyCtx()`. No more per-page re-fetching of the same
topology — the previous monolith did it implicitly by sharing
top-level state, which doesn't translate when tabs become routes.

Migration / compat:

- Legacy `/settings#discovery`, `#sources`, `#mapping`, `#resolved`
  redirect to the new subroutes via an onMount check in the layout
  so bookmarks and external links keep working.
- `#general` collapses to the (now small) `/settings` itself.
- Outbound links in NodeMappingModal and TopologySettings drawer
  updated to the new paths.

The old monolith dropped from 2779 lines to a focused 305-line page
(display settings + delete only).
